新增查询与我分享脚本功能列表接口 新增按分享人和分享时间，查询满足条件的分享脚本 新增查询分享脚本详情 新增查询拥有数据开发平台权限的所有人员信息接口 生产环境、开发环境、测试环境，添加分享功能各接口白名单

+++ geometry-mlsql-persistence/src/main/java/com/geominfo/mlsql/dao/TSystemResourcesDao.java
@@ -59,7 +59,7 @@ public interface TSystemResourcesDao extends BaseMapper<TSystemResources> {
      * @param userId
      * @return java.util.List<com.geominfo.mlsql.domain.vo.WorkSpaceInfoVo>
      */
-    List<WorkSpaceInfoResult> getWorkSpaceLists(BigDecimal userId);
+    List<WorkSpaceInfoResult> getWorkSpaceLists(Integer userId);
 
     /**
      * @description: 根据名称查询工作空间列表
@@ -79,5 +79,5 @@ public interface TSystemResourcesDao extends BaseMapper<TSystemResources> {
      * @param userId
      * @return java.util.List<com.geominfo.mlsql.domain.po.TSystemResources>
      */
-    List<TSystemResources> getRecentlyScripts(BigDecimal userId);
+    List<TSystemResources> getRecentlyScripts(Integer userId);
 }
